Changed Integer getDt to float

+++ app/src/main/java/pl/nauka/jarek/weather/model/forecast/List.java
@@ -69,7 +69,7 @@ public class List {
         this.snow = snow;
     }
 
-    public Integer getDt() {
+    public float getDt() {
         return dt;
     }
